# packages/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: CryptoGate Phase 1 API
  version: 0.3.3
  description: >
    Contract freeze v0.3.3 (platform fee tiers + merchant commercial settings).
    Builds on v0.3.2 audit + service bill lifecycle. Source of truth for apps/api,
    apps/payment-page, and apps/cashier-apk. Payment orders are watch-only.
    Service bills are a separate rail — never merge with /orders. Enums match
    @cryptogate/domain. Never return spend keys, full xPub, webhook signingSecret,
    or API key secret on GET.

    Machine clients (security apiKey): send X-Api-Key, X-Timestamp, X-Nonce,
    X-Signature. Canonical string is timestamp LF nonce LF METHOD LF pathAndQuery
    LF sha256Hex(rawBody); signature is hex HMAC-SHA256(apiSecret, canonical).
    Session cookie and public GET /orders/{id}/payment do not sign. Replay and
    skew codes: signature_invalid, timestamp_skew, nonce_replay (401). Rate
    limits: RateLimitPerMinute in domain; 429 rate_limited + Retry-After.
servers:
  - url: /v1
...
paths:
...
  /service-bills:
    get:
      tags: [service-bills]
      operationId: listServiceBills
      summary: List service bills (not payment orders)
      description: >
        Separate rail from /orders. USD subscription + volume fee. Cashier 403.
        Merchant Owner/Admin/Viewer — own org. Agent — subtree read-only.
        Platform Owner/Admin — any org via query orgId.
      security:
        - sessionCookie: []
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/apiTimestamp"
        - $ref: "#/components/parameters/apiNonce"
        - $ref: "#/components/parameters/apiSignature"
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ServiceBillStatus"
      responses:
        "200":
          description: Service bill list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceBillList"
        "403":
          description: Forbidden (e.g. Cashier)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [service-bills]
      operationId: issueServiceBill
      summary: Issue a service bill (platform only)
      description: >
        Platform Owner/Administrator only. Does not create a payment order and
        must not debit merchant settlement or payer on-chain amounts.
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueServiceBillRequest"
      responses:
        "201":
          description: Bill issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceBill"
        "403":
          description: Not a platform operator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: cg_session
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: >
        Machine client id. Must be paired with X-Timestamp, X-Nonce, X-Signature
        (see info description and M3-01). Session cookie callers omit all four.
  parameters:
...
    apiTimestamp:
      name: X-Timestamp
      in: header
      required: false
      description: >
        Unix seconds (UTC). Required when using X-Api-Key. Rejected when
        |now − value| > 300 (API_SIGNING_MAX_SKEW_SECONDS). Omit with session cookie.
      schema:
        type: string
        pattern: "^[0-9]{10,12}$"
    apiNonce:
      name: X-Nonce
      in: header
      required: false
      description: >
        16–64 chars [A-Za-z0-9_-]. Unique per API key for 600s
        (API_SIGNING_NONCE_TTL_SECONDS). Reuse → 401 nonce_replay.
      schema:
        type: string
        minLength: 16
        maxLength: 64
        pattern: "^[A-Za-z0-9_-]{16,64}$"
    apiSignature:
      name: X-Signature
      in: header
      required: false
      description: >
        Lowercase hex HMAC-SHA256(apiSecret, canonical). Canonical =
        timestamp LF nonce LF METHOD LF pathAndQuery LF sha256Hex(rawBody).
      schema:
        type: string
        pattern: "^[0-9a-f]{64}$"
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    ServiceBillStatus:
      type: string
      description: Matches @cryptogate/domain ServiceBillStatus
      enum: [issued, paid, overdue, voided]
    ServiceBill:
      type: object
      required:
        - id
        - orgId
        - periodStart
        - periodEnd
        - subscriptionAmount
        - volumeFeeAmount
        - totalAmount
        - currency
        - status
        - dueAt
      description: Platform SaaS invoice — not a payment order
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        subscriptionAmount:
          type: string
          example: "49.00"
        volumeFeeAmount:
          type: string
          example: "12.50"
        totalAmount:
          type: string
          example: "61.50"
        currency:
          type: string
          enum: [USD]
        status:
          $ref: "#/components/schemas/ServiceBillStatus"
        dueAt:
          type: string
          format: date-time
        paidAt:
          type: string
          format: date-time
          nullable: true
          description: Set when status is paid (v0.3.2)
        voidedAt:
          type: string
          format: date-time
          nullable: true
          description: Set when status is voided (v0.3.2)
        lastAdjustmentReason:
          type: string
          nullable: true
          description: Most recent platform adjustment note (v0.3.2)
    ServiceBillList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ServiceBill"
    IssueServiceBillRequest:
      type: object
      required: [orgId, periodStart, periodEnd, subscriptionAmount, volumeFeeAmount, dueAt]
      properties:
        orgId:
          type: string
          format: uuid
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        subscriptionAmount:
          type: string
        volumeFeeAmount:
          type: string
        dueAt:
          type: string
          format: date-time
...